+++ packages/wisewidgetslibrary/lib/src/widgets/widgets.dart
@@ -1,6 +1,7 @@
 export 'animated_clip_rect.dart';
 export 'animated_numbers.dart';
 export 'animated_numbers_mapped.dart';
+export 'bi_directional_paged_list.dart';
 export 'custom_image_carousel.dart';
 export 'fullscreen_image.dart';
 export 'fullscreen_image_carousel.dart';